// Backend/migrations/snapshots/ee130832d76786f06f6b670f6c69429ccd263eb465391f75d816196b026a6afc/contract.d.ts
export type StorageHash =
  StorageHashBase<'ee130832d76786f06f6b670f6c69429ccd263eb465391f75d816196b026a6afc'>;
export type ExecutionHash = ExecutionHashBase<string>;
export type ProfileHash =
  ProfileHashBase<'3916f444a8a17ad749191acf9e08dad97d1a327b88c2f1d45d12f240296aa8b2'>;

export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgTargetQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly crop: {
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/int8@1']['output'];
      readonly name: Varchar<255>;
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly crop: {
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/int8@1']['input'];
      readonly name: CodecTypes['sql/varchar@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly crop: {
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/int8@1']['output'];
      readonly name: Varchar<255>;
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly crop: {
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/int8@1']['input'];
      readonly name: CodecTypes['sql/varchar@1']['input'];
    };
  };
};

export namespace Models {
  export type public_crop = {
    description: CodecTypes['pg/text@1']['output'] | null;
    id: CodecTypes['pg/int8@1']['output'];
    name: Varchar<255>;
    readonly [RelationKeys]?: never;
  };
}

export declare const models: {
  public: {
    crop: Models.public_crop;
  };
};

export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;

type ContractBase = Omit<
  ContractType<{
    readonly namespaces: {
      readonly public: {
        readonly id: 'public';
        readonly kind: 'postgres-schema';
        readonly entries: {
          readonly table: {
            readonly crop: {
              columns: {
                readonly description: {
                  readonly nativeType: 'text';
                  readonly codecId: 'pg/text@1';
                  readonly nullable: true;
                };
                readonly id: {
                  readonly nativeType: 'int8';
                  readonly codecId: 'pg/int8@1';
                  readonly nullable: false;
                  readonly default: {
                    readonly kind: 'function';
                    readonly expression: 'autoincrement()';
                  };
                };
                readonly name: {
                  readonly nativeType: 'character varying';
                  readonly codecId: 'sql/varchar@1';
                  readonly nullable: false;
                  readonly typeParams: { readonly length: 255 };
                };
              };
              primaryKey: { readonly columns: readonly ['id'] };
              uniques: readonly [];
              indexes: readonly [];
              foreignKeys: readonly [];
            };
          };
        };
      };
    };
    readonly storageHash: StorageHash;
  }>,
  'roots' | 'domain'
> & {
  readonly target: 'postgres';
  readonly targetFamily: 'sql';
  readonly roots: {
    readonly crop: { readonly namespace: 'public' & NamespaceId; readonly model: 'crop' };
  };
  readonly domain: {
    readonly namespaces: {
      readonly public: {
        readonly models: {
          readonly crop: {
            readonly fields: {
              readonly description: {
                readonly nullable: true;
                readonly type: { readonly kind: 'scalar'; readonly codecId: 'pg/text@1' };
              };
              readonly id: {
                readonly nullable: false;
                readonly type: { readonly kind: 'scalar'; readonly codecId: 'pg/int8@1' };
              };
              readonly name: {
                readonly nullable: false;
                readonly type: {
                  readonly kind: 'scalar';
                  readonly codecId: 'sql/varchar@1';
                  readonly typeParams: { readonly length: 255 };
                };
              };
            };
            readonly relations: Record<string, never>;
            readonly storage: {
              readonly table: 'crop';
              readonly namespaceId: 'public';
              readonly fields: {
                readonly description: { readonly column: 'description' };
                readonly id: { readonly column: 'id' };
                readonly name: { readonly column: 'name' };
              };
            };
          };
        };
      };
    };
  };
  readonly capabilities: {
    readonly postgres: {
      readonly distinctOn: true;
      readonly jsonAgg: true;
      readonly lateral: true;
      readonly limit: true;
      readonly orderBy: true;
      readonly returning: true;
    };
    readonly sql: {
      readonly checkConstraint: true;
      readonly defaultInInsert: true;
      readonly enums: true;
      readonly insertOnConflictSkip: true;
      readonly insertOnConflictWithoutTarget: true;
      readonly lateral: true;
      readonly returning: true;
      readonly scalarList: true;
    };
  };
  readonly extensions: {};
  readonly meta: {};

  readonly profileHash: ProfileHash;
};

export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;